backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
supabase>=2.10.0
llama-index>=0.11.0
llama-index-vector-stores-supabase>=0.3.0
llama-index-llms-gemini>=0.3.0
llama-index-embeddings-gemini>=0.2.0
llama-index-agent-openai>=0.3.0
google-generativeai>=0.8.0
vecs>=0.4.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
